+++ omnibenchmark/io/S3versioning.py
@@ -1,14 +1,20 @@
 from itertools import groupby
-from typing import Dict
+from typing import Dict, TYPE_CHECKING
 
-import minio
+if TYPE_CHECKING:
+    import minio
+else:
+    try:
+        import minio
+    except ImportError:
+        minio = None
 
 from omnibenchmark.io.exception import MinIOStorageBucketManipulationException
 from omnibenchmark.io.RemoteStorage import is_valid_version
 
 
 def get_s3_object_versions_and_tags(
-    client: minio.Minio, benchmark: str, readonly: bool = False
+    client: "minio.Minio", benchmark: str, readonly: bool = False
 ) -> Dict:
     """
     Retrieve the metadata of all objects in a S3 Bucket.